Add explicit casts, replace constant L suffix by U, fix a few variable hiding
git-svn-id: https://svn.osgeo.org/gdal/trunk@32234 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/port/cpl_string.cpp
@@ -977,7 +977,7 @@ const char *CPLSPrintf(const char *fmt, ...)
 /*      time.                                                           */
 /* -------------------------------------------------------------------- */
     int *pnBufIndex = reinterpret_cast<int *>( pachBufRingInfo );
-    const int nOffset = sizeof(int) + *pnBufIndex * CPLSPrintf_BUF_SIZE;
+    const size_t nOffset = sizeof(int) + *pnBufIndex * CPLSPrintf_BUF_SIZE;
     char *pachBuffer = pachBufRingInfo + nOffset;
 
     *pnBufIndex = (*pnBufIndex + 1) % CPLSPrintf_BUF_Count;
